+++ interfaces/Prolog/XSB/xsb_cfli.hh
@@ -314,7 +314,7 @@ Prolog_get_atom_name(Prolog_term_ref t, Prolog_atom* ap) {
   The behavior is undefined if \p t is not a Prolog compound term.
 */
 inline int
-Prolog_get_compound_name_arity(Prolog_term_ref t, Prolog_atom* ap, int* ip) {
+Prolog_get_compound_name_arity(Prolog_term_ref t, Prolog_atom* ap, size_t* ip) {
   assert(Prolog_is_compound(t));
   *ap = p2c_functor(t);
   *ip = p2c_arity(t);
